apps/backend/requirements.txt
fastapi==0.120.0
uvicorn[standard]==0.38.0
pydantic==2.12.3
python-dotenv==1.1.1
langchain==0.3.27
langchain-openai==0.3.35
langchain-anthropic==0.3.22
langchain-google-genai==2.1.12
langgraph==0.6.11
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
